+++ client/src/entities/interview_question/api/queries.ts
@@ -2,13 +2,13 @@ import { useQuery } from '@tanstack/react-query';
 
 import { QuestionType } from '@/src/entities/interview_question/model/type';
 
-import useQuestionStore from '@/src/entities/interview_question/model/useQuestionStore';
+import useIntvPlaylistStore from '@/src/entities/interview_question/model/useIntvPlaylistStore';
 import interview from './queryKeys';
 import { getInterViewQuestionList } from './apis';
 
 /** 선택된 면접 질문 리스트 조회 */
 const useFetchQuestions = () => {
-  const { selectedType } = useQuestionStore();
+  const { selectedType } = useIntvPlaylistStore();
   return useQuery({
     queryKey: interview.list(selectedType),
     queryFn: () => getInterViewQuestionList(selectedType),
