+++ client/src/entities/interview_question/index.ts
@@ -4,5 +4,5 @@ export * as interviewQueries from './api/queries';
 export { default as queryKeys } from './api/queryKeys';
 
 export * as interviewType from './model/type';
-export { default as useQuestionStore } from './model/useQuestionStore';
+export { default as useIntvPlaylistStore } from './model/useIntvPlaylistStore';
 export { default as useRecordInfoStore } from './model/useRecordInfoStore';
