+++ client/src/entities/interview_question/model/useIntvPlaylistStore.ts
@@ -15,13 +15,11 @@ type Action = {
 /** 면접 녹화에 필요한 질문 설정 관련 정보
  *  - usePlayList : 유저가 선택한 타입별 질문
  *  - selectedTypes : 유저가 선택한 타입
- *  - confirmQuestions :  유저가 최종적으로 컨펌한 질문
  *  - initUserPlayList : 유저가 선택한 타입별 질문 초기화
  *  - initConfromQuestions : 유저가 면접리스트로 컨펌한 질문 초기화
 */
-const useQuestionStore = create<State & Action>((set) => ({
+const useIntvPlaylistStore = create<State & Action>((set) => ({
   userPlayList: [],
-  confirmQuestions: [],
   selectedType: '',
 
   setUserPlayList: (list: string[]) =>
@@ -47,4 +45,4 @@ const useQuestionStore = create<State & Action>((set) => ({
   initUserPlayList: () => set({ userPlayList: [] }),
 }));
 
-export default useQuestionStore;
+export default useIntvPlaylistStore;
