+++ client/src/features/interview/expected-question/ui/selector/QuestionTypeSelector.tsx
@@ -3,7 +3,7 @@ import { ComboBox } from '@/src/shared/ui/combobox';
 
 import {
   interviewType,
-  useQuestionStore,
+  useIntvPlaylistStore,
 } from '@/src/entities/interview_question';
 import { V } from '@/src/shared/styles';
 
@@ -24,7 +24,7 @@ interface IQuestionSelectorProps {
  * @param onReset : 질문 체크 해제
  */
 function QuestionTypeSelector({ onReset }: IQuestionSelectorProps) {
-  const { selectedType, setSelectedType } = useQuestionStore();
+  const { selectedType, setSelectedType } = useIntvPlaylistStore();
   const onChangeType = (type: interviewType.QuestionType | string) => {
     onReset();
     setSelectedType(type);
